Split DataProvidersTest.php into two folders
